backend/pyproject.toml
[project]
name = "repl-nix-new-repl-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "boto3>=1.35.82",
    "fastapi>=0.115.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]
